Added support for direct routes to containers via dnsdock.
Removed dnsmasq and HTTP proxy support as they're no longer required.
Also added command to retrieve nameserver IP.
