+++ cli/dinghy/dnsdock.rb
@@ -1,0 +1,70 @@
+require 'stringio'
+require 'tempfile'
+
+require 'dinghy/machine'
+
+class Dnsdock
+  CONTAINER_NAME = "dinghy_dnsdock"
+  CONTAINER_IP = "172.17.42.1"
+  RESOLVER_DIR = Pathname("/etc/resolver")
+  RESOLVER_FILE = RESOLVER_DIR.join("docker")
+
+  attr_reader :machine
+
+  def initialize(machine)
+    @machine = machine
+  end
+
+  def up
+    puts "Starting the DNS nameserver"
+    System.capture_output do
+      docker.system("rm", "-fv", CONTAINER_NAME)
+    end
+    docker.system("run", "-d", "-v", "/var/run/docker.sock:/var/run/docker.sock", "--name", CONTAINER_NAME, "-p", "#{CONTAINER_IP}:53:53/udp", "tonistiigi/dnsdock", '-domain="docker"')
+  end
+
+  def status
+    return "not running" if !machine.running?
+
+    output, _ = System.capture_output do
+      docker.system("inspect", "-f", "{{ .State.Running }}", CONTAINER_NAME)
+    end
+
+    if output.strip == "true"
+      "running"
+    else
+      "not running"
+    end
+  end
+
+  def configure_resolver!
+    puts "setting up DNS resolution, this will require sudo"
+    unless RESOLVER_DIR.directory?
+      system!("creating #{RESOLVER_DIR}", "sudo", "mkdir", "-p", RESOLVER_DIR)
+    end
+    Tempfile.open('dinghy-dnsdock') do |f|
+      f.write(resolver_contents)
+      f.close
+      system!("creating #{RESOLVER_FILE}", "sudo", "cp", f.path, RESOLVER_FILE)
+      system!("creating #{RESOLVER_FILE}", "sudo", "chmod", "644", RESOLVER_FILE)
+    end
+    system!("creating route", "sudo", "route", "-n", "add", "172.17.0.0/16", machine.vm_ip)
+    system!("restarting mDNSResponder", "sudo", "killall", "mDNSResponder")
+  end
+
+  def resolver_configured?
+    RESOLVER_FILE.exist? && File.read(RESOLVER_FILE) == resolver_contents
+  end
+
+  def resolver_contents; <<-EOS.gsub(/^    /, '')
+    # Generated by dinghy
+    nameserver #{CONTAINER_IP}
+    EOS
+  end
+
+  private
+
+  def docker
+    @docker ||= Docker.new(machine)
+  end
+end
